fix 2

+++ conversions/Forwards/src/main/java/com/agentdid127/resourcepack/forwards/impl/NameConverter.java
@@ -84,10 +84,10 @@ public class NameConverter extends Converter {
                 //Update 1.14 items
                 if (to > Util.getVersionProtocol(packConverter.getGson(), "1.13")) {
                     renameAll(newItemMapping, ".json", models.resolve("item"));
-                    Files.copy(models.resolve("item" + File.separator + "black_dye.json"), models.resolve("item" + File.separator + "ink_sac.json"));
-                    Files.copy(models.resolve("item" + File.separator + "brown_dye.json"), models.resolve("item" + File.separator + "cocoa_beans.json"));
-                    Files.copy(models.resolve("item" + File.separator + "white_dye.json"), models.resolve("item" + File.separator + "bone_meal.json"));
-                    Files.copy(models.resolve("item" + File.separator + "blue_dye.json"), models.resolve("item" + File.separator + "lapis_lazuli.json"));
+                    if (!models.resolve("item" + File.separator + "ink_sac.json").toFile().exists()) Files.copy(models.resolve("item" + File.separator + "black_dye.json"), models.resolve("item" + File.separator + "ink_sac.json"));
+                    if (!models.resolve("item" + File.separator + "cocoa_beans.json").toFile().exists()) Files.copy(models.resolve("item" + File.separator + "brown_dye.json"), models.resolve("item" + File.separator + "cocoa_beans.json"));
+                    if (!models.resolve("item" + File.separator + "bone_meal.json").toFile().exists())Files.copy(models.resolve("item" + File.separator + "white_dye.json"), models.resolve("item" + File.separator + "bone_meal.json"));
+                    if (!models.resolve("item" + File.separator + "lapis_lazuli.json").toFile().exists())Files.copy(models.resolve("item" + File.separator + "blue_dye.json"), models.resolve("item" + File.separator + "lapis_lazuli.json"));
                 }
 
                 if (to > Util.getVersionProtocol(packConverter.getGson(), "1.19")) {
